Fix profile builder

+++ src/fbuilder/build_profile.c
@@ -61,19 +61,18 @@ void build_profile(int argc, char **argv, int index, FILE *fp) {
 
 	char *output;
 	char *stroutput;
-	if(asprintf(&output,"--output=%s",trace_output) == -1)
+	if(asprintf(&output,"--trace=%s",trace_output) == -1)
 		errExit("asprintf");
-	if(asprintf(&stroutput,"-o %s",strace_output) == -1)
+	if(asprintf(&stroutput,"-o%s",strace_output) == -1)
 		errExit("asprintf");
 
 	char *cmdlist[] = {
-	  "/usr/bin/firejail",
+	  BINDIR "/firejail",
 	  "--quiet",
-	  output,
 	  "--noprofile",
 	  "--caps.drop=all",
 	  "--nonewprivs",
-	  "--trace",
+	  output,
 	  "--shell=none",
 	  "/usr/bin/strace", // also used as a marker in build_profile()
 	  "-c",
@@ -110,7 +109,7 @@ void build_profile(int argc, char **argv, int index, FILE *fp) {
 
 	if (arg_debug) {
 		for (i = 0; i < len; i++)
-			printf("\t%s\n", cmd[i]);
+			printf("%s%s\n", (i)?"\t":"", cmd[i]);
 	}
 
 	// fork and execute
@@ -130,7 +129,8 @@ void build_profile(int argc, char **argv, int index, FILE *fp) {
 		errExit("waitpid");
 
 	if (WIFEXITED(status) && WEXITSTATUS(status) == 0) {
-		printf("\n\n\n");
+		if (fp == stdout)
+			printf("--- Built profile beings after this line ---\n");
 		fprintf(fp, "############################################\n");
 		fprintf(fp, "# %s profile\n", argv[index]);
 		fprintf(fp, "############################################\n");
@@ -177,9 +177,10 @@ void build_profile(int argc, char **argv, int index, FILE *fp) {
 		fprintf(fp, "### environment\n");
 		fprintf(fp, "shell none\n");
 
-		unlink(trace_output);
-		unlink(strace_output);
-
+		if (!arg_debug) {
+			unlink(trace_output);
+			unlink(strace_output);
+		}
 	}
 	else {
 		fprintf(stderr, "Error: cannot run the sandbox\n");
